refactor: move persisted transaction type to correct file
