+++ packages/shared/src/lib/core/transactions/types/index.ts
@@ -1,0 +1,1 @@
+export * from './persisted-transaction.type'
