+++ packages/shared/src/lib/core/transactions/types/persisted-transaction.type.ts
@@ -1,0 +1,16 @@
+import { IBlockscoutTransaction } from '@auxiliary/blockscout/interfaces'
+import { PersistedEvmTransaction } from '@core/activity'
+
+export type PersistedTransaction =
+    | {
+          blockscout: IBlockscoutTransaction
+          local: PersistedEvmTransaction
+      }
+    | {
+          blockscout?: IBlockscoutTransaction
+          local: PersistedEvmTransaction
+      }
+    | {
+          blockscout: IBlockscoutTransaction
+          local?: PersistedEvmTransaction
+      }
